EXPERIMENTAL: Support for async writes for buffered files
This adds support for QD=2 writes for buffered (e.g WAL) writes.
This improves write tail latencies quite a bit and brings
us close to the tail latencies of regular file system writes.

Currently experimental, needs more work and validation before merging
to master.

This change also adds a dependency to libaio, which is not really needed,
we could work with the kernel ABI directly.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>